Protecting user notes with auth middleware
Protecting user notes with auth middleware

+++ backend/src/controllers/notes.ts
@@ -2,11 +2,17 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes:RequestHandler = async (req, res, next)=>{
+    const authenticatedUserId=req.session.userId;
+    
+    
     try {
-        //throw createHttpError(401);
-        const notes = await NoteModel.find().exec();
+        
+        assertIsDefined(authenticatedUserId);
+
+        const notes = await NoteModel.find({userId: authenticatedUserId}).exec();
         res.status(200).json(notes);
     } catch (error) {
         next(error);
@@ -15,8 +21,11 @@ export const getNotes:RequestHandler = async (req, res, next)=>{
 
 export const getNote:RequestHandler = async (req, res, next)=>{
     const noteId=req.params.noteId;
+    const authenticatedUserId=req.session.userId;
     
     try {
+
+        assertIsDefined(authenticatedUserId);
 
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400,"Invalid Note ID!");
@@ -26,6 +35,10 @@ export const getNote:RequestHandler = async (req, res, next)=>{
         
         if(!note){
             throw createHttpError(404,"Note not Found!");
+        }
+
+        if(!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401,"You cannot access this note.");
         }
         
         res.status(200).json(note);
@@ -47,12 +60,16 @@ export const createNote: RequestHandler<unknown, unknown, CreateNoteBody, unknow
     
     const title = req.body.title;
     const text = req.body.text;
+    const authenticatedUserId=req.session.userId;
     
     try {
+        assertIsDefined(authenticatedUserId);
+
         if(!title){
             throw createHttpError(400,"Note must have a title!");
         }
         const newNote=await NoteModel.create({
+            userId: authenticatedUserId,
             title: title,
             text: text,
         });
@@ -77,7 +94,9 @@ export const updateNote:RequestHandler<UpdateNoteParams, unknown, updateNoteBody
     const noteId=req.params.noteId;
     const newTitle = req.body.title;
     const newText = req.body.text;
+    const authenticatedUserId=req.session.userId;
     try {
+        assertIsDefined(authenticatedUserId);
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400,"Invalid Note ID!");
         }
@@ -88,6 +107,10 @@ export const updateNote:RequestHandler<UpdateNoteParams, unknown, updateNoteBody
         const note = await NoteModel.findById(noteId).exec();
         if(!note){
             throw createHttpError(404,"Note not Found!");
+        }
+
+        if(!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401,"You cannot access this note.");
         }
 
         note.title=newTitle;
@@ -104,7 +127,9 @@ export const updateNote:RequestHandler<UpdateNoteParams, unknown, updateNoteBody
 
 export const deleteNote: RequestHandler = async(req,res,next)=>{
     const noteId=req.params.noteId;
+    const authenticatedUserId=req.session.userId;
     try {
+        assertIsDefined(authenticatedUserId);
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400,"Invalid note id.");
         }
@@ -113,6 +138,10 @@ export const deleteNote: RequestHandler = async(req,res,next)=>{
         if(!note){
             throw createHttpError(404,"Note Not Found!");
         }
+        if(!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401,"You cannot access this note.");
+        }
+
 
         await note.deleteOne();
